feat: add infocards and bar chart
OK! I did too much here 🤦🏾‍♂️
I know that from the beginning that D3 is hard, being a library that
provides low-level customizations, like Tailwind CSS. Like any other
thing in tech that provides this level of customizations, the learning
curve is steep. Oh well, I just take it as an exercise with the belief
that I'll get back at it with time. After all, I used two months to
learn how to use functions in JavaScript 😅

As for the UI, I see why frontend devs don't like the idea of using a
component library, especially when you are building a product for a
startup, putting Tailwind CSS in a sweet spot. Well, if you ask me, use
a UI library only if you want structure in building your web interfaces.

So yeah.
That's just it.

Till the next commit 👋