style!: added colour scheme
So with the help of trusty ChatGPT, I was able to create a colour scheme
for the application. It will ease out picking colours for different
components and interactions.

Then, I finally got the hang of the D3 chart system. It took two days of
practicing and tweaking, but it was worth it.

From here, it is to add the rest of the necessary components in order to
build the 'Overview' page of the application.

So yeah.
That's it for now.

Till the next commit 👋